pyproject.toml
[project]
name = "projecttemplatev2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arrow==1.4.0",
    "asgiref==3.11.1",
    "azure-core==1.38.0",
    "azure-identity==1.25.1",
    "azure-keyvault-secrets==4.10.0",
    "binaryornot==0.4.4",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "chardet==5.2.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cookiecutter==2.6.0",
    "cryptography==46.0.4",
    "diff-match-patch==20241021",
    "django==6.0.2",
    "django-allauth==65.14.0",
    "django-auditlog==3.4.1",
    "django-debug-toolbar==6.2.0",
    "django-extensions==4.1",
    "django-filter==25.2",
    "django-import-export==4.4.0",
    "django-tailwind==4.4.2",
    "django-unfold==0.78.1",
    "dotenv>=0.9.9",
    "gunicorn==25.0.1",
    "honcho==2.0.0",
    "idna==3.11",
    "isodate==0.7.2",
    "jinja2==3.1.6",
    "loguru==0.7.3",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "msal==1.34.0",
    "msal-extensions==1.3.1",
    "packaging==26.0",
    "pillow==12.1.0",
    "psycopg-binary==3.3.2",
    "psycopg[binary]==3.3.2",
    "pycparser==3.0",
    "pygments==2.19.2",
    "pyjwt==2.11.0",
    "pyopenssl==25.3.0",
    "pytailwindcss==0.3.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-slugify==8.0.4",
    "python3-saml==1.16.0",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rich==14.3.2",
    "six==1.17.0",
    "sqlparse==0.5.5",
    "tablib==3.9.0",
    "text-unidecode==1.3",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "werkzeug==3.1.5",
    "xmlsec==1.3.17",
]
